int16 to 32 fix

+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/CRM/CRMUsers.ascx.cs
@@ -114,7 +114,7 @@ namespace SolidCP.Portal.CRM
 
         protected void ddlPageSize_SelectedIndexChanged(object sender, EventArgs e)
         {
-            gvUsers.PageSize = Convert.ToInt16(ddlPageSize.SelectedValue);
+            gvUsers.PageSize = Convert.ToInt32(ddlPageSize.SelectedValue);
 
             // rebind grid
             gvUsers.DataBind();
