+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/ExchangeServer/ExchangeContacts.ascx.cs
@@ -92,7 +92,7 @@ namespace SolidCP.Portal.ExchangeServer
 
         protected void ddlPageSize_SelectedIndexChanged(object sender, EventArgs e)   
         {   
-            gvContacts.PageSize = Convert.ToInt16(ddlPageSize.SelectedValue);   
+            gvContacts.PageSize = Convert.ToInt32(ddlPageSize.SelectedValue);   
 
             // rebind grid   
             gvContacts.DataBind();
