+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/ExchangeServer/ExchangeMailboxes.ascx.cs
@@ -211,7 +211,7 @@ namespace SolidCP.Portal.ExchangeServer
 
         protected void ddlPageSize_SelectedIndexChanged(object sender, EventArgs e)   
         {   
-            gvMailboxes.PageSize = Convert.ToInt16(ddlPageSize.SelectedValue);   
+            gvMailboxes.PageSize = Convert.ToInt32(ddlPageSize.SelectedValue);   
        
             // rebind grid   
             gvMailboxes.DataBind();
